+++ ccc-20200701/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("d4c00ed0-66f2-11ef-98c1-51bae8c6d9f1")]
+[assembly: Guid("fd6a8840-6cc1-11ef-b7f4-3fa7331034bf")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("2.29.1.0")]
-[assembly: AssemblyFileVersion("2.29.1.0")]
+[assembly: AssemblyVersion("2.29.2.0")]
+[assembly: AssemblyFileVersion("2.29.2.0")]